ai around the hre will no longer join
did you know that there is no way to mod out joining the hre? did you know that there is also no way to mod the requirements for joining the hre? so now we have this ghetto ass negative opinion modifier shit with austria so ai around the hre doesn't join.